extract common workload structs and tests to workload.go
