+++ internal/pkg/manifest/svc.go
@@ -5,9 +5,7 @@
 package manifest
 
 import (
-	"errors"
 	"fmt"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
@@ -22,17 +20,7 @@ const (
 	LoadBalancedWebServiceType = "Load Balanced Web Service"
 	// BackendServiceType is a service that cannot be accessed from the internet but can be reached from other services.
 	BackendServiceType = "Backend Service"
-
-	defaultSidecarPort    = "80"
-	defaultFluentbitImage = "amazon/aws-for-fluent-bit:latest"
 )
-
-var (
-	errUnmarshalBuildOpts = errors.New("can't unmarshal build field into string or compose-style map")
-	errUnmarshalCountOpts = errors.New(`unmarshal "count" field to an integer or autoscaling configuration`)
-)
-
-var dockerfileDefaultName = "Dockerfile"
 
 // ServiceTypes are the supported service manifest types.
 var ServiceTypes = []string{
@@ -61,210 +49,10 @@ func (r Range) Parse() (min int, max int, err error) {
 	return min, max, nil
 }
 
-// Workload holds the basic data that every workload manifest file needs to have.
-type Workload struct {
-	Name *string `yaml:"name"`
-	Type *string `yaml:"type"` // must be one of the supported manifest types.
-}
-
-// Image represents the workload's container image.
-type Image struct {
-	Build BuildArgsOrString `yaml:"build"` // Path to the Dockerfile.
-}
-
-// BuildConfig populates a docker.BuildArguments struct from the fields available in the manifest.
-// Prefer the following hierarchy:
-// 1. Specific dockerfile, specific context
-// 2. Specific dockerfile, context = dockerfile dir
-// 3. "Dockerfile" located in context dir
-// 4. "Dockerfile" located in ws root.
-func (i *Image) BuildConfig(rootDirectory string) *DockerBuildArgs {
-	df := i.dockerfile()
-	ctx := i.context()
-	if df != "" && ctx != "" {
-		return &DockerBuildArgs{
-			Dockerfile: aws.String(filepath.Join(rootDirectory, df)),
-			Context:    aws.String(filepath.Join(rootDirectory, ctx)),
-			Args:       i.args(),
-		}
-	}
-	if df != "" && ctx == "" {
-		return &DockerBuildArgs{
-			Dockerfile: aws.String(filepath.Join(rootDirectory, df)),
-			Context:    aws.String(filepath.Join(rootDirectory, filepath.Dir(df))),
-			Args:       i.args(),
-		}
-	}
-	if df == "" && ctx != "" {
-		return &DockerBuildArgs{
-			Dockerfile: aws.String(filepath.Join(rootDirectory, ctx, dockerfileDefaultName)),
-			Context:    aws.String(filepath.Join(rootDirectory, ctx)),
-			Args:       i.args(),
-		}
-	}
-	return &DockerBuildArgs{
-		Dockerfile: aws.String(filepath.Join(rootDirectory, dockerfileDefaultName)),
-		Context:    aws.String(rootDirectory),
-		Args:       i.args(),
-	}
-}
-
-// dockerfile returns the path to the workload's Dockerfile. If no dockerfile is specified,
-// returns "".
-func (i *Image) dockerfile() string {
-	// Prefer to use the "Dockerfile" string in BuildArgs. Otherwise,
-	// "BuildString". If no dockerfile specified, return "".
-	if i.Build.BuildArgs.Dockerfile != nil {
-		return aws.StringValue(i.Build.BuildArgs.Dockerfile)
-	}
-
-	var dfPath string
-	if i.Build.BuildString != nil {
-		dfPath = aws.StringValue(i.Build.BuildString)
-	}
-
-	return dfPath
-}
-
-// context returns the build context directory if it exists, otherwise an empty string.
-func (i *Image) context() string {
-	return aws.StringValue(i.Build.BuildArgs.Context)
-}
-
-// args returns the args section, if it exists, to override args in the dockerfile.
-// Otherwise it returns an empty map.
-func (i *Image) args() map[string]string {
-	return i.Build.BuildArgs.Args
-}
-
-// BuildArgsOrString is a custom type which supports unmarshaling yaml which
-// can either be of type string or type DockerBuildArgs.
-type BuildArgsOrString struct {
-	BuildString *string
-	BuildArgs   DockerBuildArgs
-}
-
-// UnmarshalYAML overrides the default YAML unmarshaling logic for the BuildArgsOrString
-// struct, allowing it to perform more complex unmarshaling behavior.
-// This method implements the yaml.Unmarshaler (v2) interface.
-func (b *BuildArgsOrString) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	if err := unmarshal(&b.BuildArgs); err != nil {
-		switch err.(type) {
-		case *yaml.TypeError:
-			break
-		default:
-			return err
-		}
-	}
-
-	if !b.BuildArgs.isEmpty() {
-		// Unmarshaled successfully to b.BuildArgs, return.
-		return nil
-	}
-
-	if err := unmarshal(&b.BuildString); err != nil {
-		return errUnmarshalBuildOpts
-	}
-	return nil
-}
-
-// DockerBuildArgs represents the options specifiable under the "build" field
-// of Docker Compose services. For more information, see:
-// https://docs.docker.com/compose/compose-file/#build
-type DockerBuildArgs struct {
-	Context    *string           `yaml:"context,omitempty"`
-	Dockerfile *string           `yaml:"dockerfile,omitempty"`
-	Args       map[string]string `yaml:"args,omitempty"`
-}
-
-func (b *DockerBuildArgs) isEmpty() bool {
-	if b.Context == nil && b.Dockerfile == nil && b.Args == nil {
-		return true
-	}
-	return false
-}
-
 // ServiceImageWithPort represents a container image with an exposed port.
 type ServiceImageWithPort struct {
 	Image `yaml:",inline"`
 	Port  *uint16 `yaml:"port"`
-}
-
-// LogConfig holds configuration for Firelens to route your logs.
-type LogConfig struct {
-	Image          *string           `yaml:"image"`
-	Destination    map[string]string `yaml:"destination,flow"`
-	EnableMetadata *bool             `yaml:"enableMetadata"`
-	SecretOptions  map[string]string `yaml:"secretOptions"`
-	ConfigFile     *string           `yaml:"configFilePath"`
-}
-
-func (lc *LogConfig) logConfigOpts() *template.LogConfigOpts {
-	return &template.LogConfigOpts{
-		Image:          lc.image(),
-		ConfigFile:     lc.ConfigFile,
-		EnableMetadata: lc.enableMetadata(),
-		Destination:    lc.Destination,
-		SecretOptions:  lc.SecretOptions,
-	}
-}
-
-func (lc *LogConfig) image() *string {
-	if lc.Image == nil {
-		return aws.String(defaultFluentbitImage)
-	}
-	return lc.Image
-}
-
-func (lc *LogConfig) enableMetadata() *string {
-	if lc.EnableMetadata == nil {
-		// Enable ecs log metadata by default.
-		return aws.String("true")
-	}
-	return aws.String(strconv.FormatBool(*lc.EnableMetadata))
-}
-
-// Sidecar holds configuration for all sidecar containers in a service.
-type Sidecar struct {
-	Sidecars map[string]*SidecarConfig `yaml:"sidecars"`
-}
-
-// Options converts the service's sidecar configuration into a format parsable by the templates pkg.
-func (s *Sidecar) Options() ([]*template.SidecarOpts, error) {
-	if s.Sidecars == nil {
-		return nil, nil
-	}
-	var sidecars []*template.SidecarOpts
-	for name, config := range s.Sidecars {
-		port, protocol, err := parsePortMapping(config.Port)
-		if err != nil {
-			return nil, err
-		}
-		sidecars = append(sidecars, &template.SidecarOpts{
-			Name:       aws.String(name),
-			Image:      config.Image,
-			Port:       port,
-			Protocol:   protocol,
-			CredsParam: config.CredsParam,
-		})
-	}
-	return sidecars, nil
-}
-
-// SidecarConfig represents the configurable options for setting up a sidecar container.
-type SidecarConfig struct {
-	Port       *string `yaml:"port"`
-	Image      *string `yaml:"image"`
-	CredsParam *string `yaml:"credentialsParameter"`
-}
-
-// TaskConfig represents the resource boundaries and environment variables for the containers in the task.
-type TaskConfig struct {
-	CPU       *int              `yaml:"cpu"`
-	Memory    *int              `yaml:"memory"`
-	Count     Count             `yaml:"count"`
-	Variables map[string]string `yaml:"variables"`
-	Secrets   map[string]string `yaml:"secrets"`
 }
 
 // Count is a custom type which supports unmarshaling yaml which
@@ -342,54 +130,6 @@ func (a *Autoscaling) IsEmpty() bool {
 		a.Requests == nil && a.ResponseTime == nil
 }
 
-// WorkloadProps contains properties for creating a new service manifest.
-type WorkloadProps struct {
-	Name       string
-	Dockerfile string
-}
-
-// UnmarshalWorkload deserializes the YAML input stream into a service manifest object.
-// If an error occurs during deserialization, then returns the error.
-// If the service type in the manifest is invalid, then returns an ErrInvalidManifestType.
-func UnmarshalWorkload(in []byte) (interface{}, error) {
-	am := Workload{}
-	if err := yaml.Unmarshal(in, &am); err != nil {
-		return nil, fmt.Errorf("unmarshal to service manifest: %w", err)
-	}
-	typeVal := aws.StringValue(am.Type)
-
-	switch typeVal {
-	case LoadBalancedWebServiceType:
-		m := newDefaultLoadBalancedWebService()
-		if err := yaml.Unmarshal(in, m); err != nil {
-			return nil, fmt.Errorf("unmarshal to load balanced web service: %w", err)
-		}
-		return m, nil
-	case BackendServiceType:
-		m := newDefaultBackendService()
-		if err := yaml.Unmarshal(in, m); err != nil {
-			return nil, fmt.Errorf("unmarshal to backend service: %w", err)
-		}
-		if m.BackendServiceConfig.Image.HealthCheck != nil {
-			// Make sure that unset fields in the healthcheck gets a default value.
-			m.BackendServiceConfig.Image.HealthCheck.applyIfNotSet(newDefaultContainerHealthCheck())
-		}
-		return m, nil
-	case ScheduledJobType:
-		m := newDefaultScheduledJob()
-		if err := yaml.Unmarshal(in, m); err != nil {
-			return nil, fmt.Errorf("unmarshal to scheduled job: %w", err)
-		}
-		return m, nil
-	default:
-		return nil, &ErrInvalidWkldManifestType{Type: typeVal}
-	}
-}
-
-func durationp(v time.Duration) *time.Duration {
-	return &v
-}
-
 // Valid sidecar portMapping example: 2000/udp, or 2000 (default to be tcp).
 func parsePortMapping(s *string) (port *string, protocol *string, err error) {
 	if s == nil {
@@ -405,4 +145,8 @@ func parsePortMapping(s *string) (port *string, protocol *string, err error) {
 	default:
 		return nil, nil, fmt.Errorf("cannot parse port mapping from %s", *s)
 	}
+}
+
+func durationp(v time.Duration) *time.Duration {
+	return &v
 }
